Add tests for derive_precursor_mz_from_parent_mass
