+++ tests/filtering/test_add_parent_mass.py
@@ -1,6 +1,5 @@
 import numpy as np
 import pytest
-from matchms.constants import PROTON_MASS
 from matchms.filtering import add_parent_mass
 from ..builder_Spectrum import SpectrumBuilder
 
@@ -27,21 +26,6 @@ def test_add_parent_mass_precursormz(caplog):
     assert np.abs(spectrum.get("parent_mass") - 445.0) < .01, "Expected parent mass of about 445.0."
     assert isinstance(spectrum.get("parent_mass"), float), "Expected parent mass to be float."
     assert "Not sufficient spectrum metadata to derive parent mass." not in caplog.text
-
-
-@pytest.mark.parametrize("adduct, expected", [("[M+2Na-H]+", 399.02884),
-                                              ("[M+H+NH4]2+", 212.47945),
-                                              ("[2M+FA-H]-", 843.001799),
-                                              ("M+H", 442.992724),
-                                              ("M+H-H2O", 461.003289)])
-def test_add_parent_mass_using_adduct(adduct, expected):
-    """Test if parent mass is correctly derived from adduct information."""
-    metadata = {"precursor_mz": 444.0, "adduct": adduct, "charge": +1}
-    spectrum_in = SpectrumBuilder().with_metadata(metadata).build()
-    spectrum = add_parent_mass(spectrum_in)
-
-    assert np.allclose(spectrum.get("parent_mass"), expected, atol=1e-4), f"Expected parent mass of about {expected}."
-    assert isinstance(spectrum.get("parent_mass"), float), "Expected parent mass to be float."
 
 
 @pytest.mark.parametrize("overwrite, expected", [(True, 442.992724),
@@ -92,17 +76,3 @@ def test_empty_spectrum():
     spectrum = add_parent_mass(spectrum_in)
 
     assert spectrum is None, "Expected different handling of None spectrum."
-
-
-@pytest.mark.parametrize("ionmode, expected", [("positive", 444.0 - PROTON_MASS),
-                                               ("negative", 444.0 + PROTON_MASS)])
-def test_use_of_ionmode(ionmode, expected):
-    """Test when there is no charge given, than the ionmode
-    is used to derive parent mass."""
-    metadata = {"precursor_mz": 444.0, "ionmode": ionmode}
-    spectrum_in = SpectrumBuilder().with_metadata(metadata).build()
-
-    spectrum = add_parent_mass(spectrum_in)
-
-    assert spectrum.get("parent_mass") == expected, \
-        "Expected a different parent_mass"
